use prebuilt openblas packages
travis-ci uses ubuntu 12.0 (precise) with the option to use ubuntu trusty.  Neither provide a version of openblas that can pass the drotmg test due to https://github.com/xianyi/OpenBLAS/issues/365, which was fixed after openblas 2.10 (I think).

I've circumvented the typical apt-get process here.  On the plus side this is much faster.  On the minus side it is more fragile, and we are not testing against the absolute latest OpenBLAS.